+++ amadeus/travel/__init__.py
@@ -1,5 +1,5 @@
 from ._analytics import Analytics
 from ._predictions import TripPurpose, FlightDelay
-from ._trip_parser_jobs import TripParser
+from ._trip_parser import TripParser
 
 __all__ = ['Analytics', 'TripPurpose', 'FlightDelay', 'TripParser']
