+++ amadeus/travel/_trip_parser.py
